+++ walkers/src/lib.rs
@@ -14,7 +14,7 @@ mod zoom;
 
 pub use download::{HeaderValue, HttpOptions};
 pub use map::{Map, MapMemory, Plugin, Projector};
-pub use mercator::{screen_to_position, TileId};
+pub use mercator::TileId;
 pub use position::{lat_lon, lon_lat, Position};
 pub use tiles::{HttpTiles, Texture, TextureWithUv, Tiles};
 pub use zoom::InvalidZoom;