+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+* `screen_to_position` is no longer a public function. Use `Projector::unproject` to obtain
+  geographical coordinates from screen coordinates relative to the map viewport.
+
 ## 0.35.0
 
 * Fixed zooming on touchscreens so that it follows the center of the touch points.